requirements.txt
django>=3.1.0,<3.2.0
djangorestframework>=3.12.0,<3.13.0
flake8>=3.8.0,<3.9.0
django-sortedm2m>=3.0.0,<3.1.0
python-decouple==3.4
sat-search==0.3.0